Applies new naming convention
