+++ plugins/.gitignore
@@ -1,6 +1,6 @@
 public/*/plugin.js
 
 # Ignore files generated by build
+auth/public/auth-oidc-pkce/*.LICENSE.txt
 auth/public/auth-oidc/*.LICENSE.txt
-auth/public/auth-oidc-legacy/*.LICENSE.txt
-auth/public/auth-oidc-legacy/oidc-client.min.js.LICENSE.txt
+auth/public/auth-oidc/oidc-client.min.js.LICENSE.txt
