--- plugins/auth/public/auth-oidc-legacy/.gitignore
@@ -1,2 +1,0 @@
-# vendor
-oidc-client.min.js
