+++ plugins/auth/public/auth-oidc-pkce/.gitignore
@@ -1,0 +1,2 @@
+# vendor
+oidc-client-ts.min.js
