+++ plugins/auth/public/auth-oidc-pkce/silent-callback.html
@@ -1,5 +1,4 @@
-<script src="/assets/auth-oidc-legacy/oidc-client.min.js"></script>
-
+<script src="/assets/auth-oidc-pkce/oidc-client-ts.min.js"></script>
 <script>
   var mgr = new Oidc.UserManager();
   mgr.signinSilentCallback().catch(error => {
