+++ plugins/auth/public/auth-oidc/.gitignore
@@ -1,2 +1,2 @@
 # vendor
-oidc-client-ts.min.js
+oidc-client.min.js
