+++ plugins/auth/public/auth-oidc/silent-callback.html
@@ -1,4 +1,5 @@
-<script src="/assets/auth-oidc/oidc-client-ts.min.js"></script>
+<script src="/assets/auth-oidc/oidc-client.min.js"></script>
+
 <script>
   var mgr = new Oidc.UserManager();
   mgr.signinSilentCallback().catch(error => {
