+++ plugins/auth/src/auth-oidc/webpack-extra.config.js
@@ -8,7 +8,7 @@ module.exports = {
     new CopyWebpackPlugin({
       patterns: [
         {
-          from: pluginRoot + 'node_modules/oidc-client-ts/dist/browser/oidc-client-ts.min.js',
+          from: pluginRoot + 'node_modules/oidc-client/dist/oidc-client.min.js',
           to: pluginRoot + 'auth/public/auth-oidc'
         }
       ]
