+++ scripts/testCompatibility.sh
@@ -196,7 +196,7 @@ linkLuigi() {
     ln -s $LUIGI_DIR/client/public $NODE_MODULES_LUIGI/client
     ln -s $LUIGI_DIR/plugins/auth/public/auth-oauth2 $NODE_MODULES_LUIGI/plugin-auth-oauth2
     ln -s $LUIGI_DIR/plugins/auth/public/auth-oidc $NODE_MODULES_LUIGI/plugin-auth-oidc
-    ln -s $LUIGI_DIR/plugins/auth/public/auth-oidc-legacy $NODE_MODULES_LUIGI/plugin-auth-oidc-legacy
+    ln -s $LUIGI_DIR/plugins/auth/public/auth-oidc-pkce $NODE_MODULES_LUIGI/plugin-auth-oidc-pkce
     ln -s $LUIGI_DIR/client-frameworks-support/client-support-angular/dist/client-support-angular $NODE_MODULES_LUIGI/client-support-angular
     ln -s $LUIGI_DIR/client-frameworks-support/testing-utilities/dist $NODE_MODULES_LUIGI/testing-utilities
 
@@ -207,7 +207,7 @@ linkLuigi() {
     ls $NODE_MODULES_LUIGI/client
     ls $NODE_MODULES_LUIGI/plugin-auth-oauth2
     ls $NODE_MODULES_LUIGI/plugin-auth-oidc
-    ls $NODE_MODULES_LUIGI/plugin-auth-oidc-legacy
+    ls $NODE_MODULES_LUIGI/plugin-auth-oidc-pkce
     ls $NODE_MODULES_LUIGI/client-support-angular
     ls $NODE_MODULES_LUIGI/testing-utilities
     echoe "Finished printing contents"
@@ -225,12 +225,12 @@ linkLuigi() {
       echoe "There was an issue linking the auth-oauth2 module"
       exit 2
     fi
-    if [ ! -f $NODE_MODULES_LUIGI/plugin-auth-oidc/package.json ]; then
-      echoe "There was an issue linking the auth-oidc module"
+    if [ ! -f $NODE_MODULES_LUIGI/plugin-auth-oidc-pkce/package.json ]; then
+      echoe "There was an issue linking the auth-oidc-pkce module"
       exit 2
     fi
-    if [ ! -f $NODE_MODULES_LUIGI/plugin-auth-oidc-legacy/package.json ]; then
-      echoe "There was an issue linking the auth-oidc-legacy module"
+    if [ ! -f $NODE_MODULES_LUIGI/plugin-auth-oidc/package.json ]; then
+      echoe "There was an issue linking the auth-oidc module"
       exit 2
     fi
     if [ ! -f $NODE_MODULES_LUIGI/client-support-angular/package.json ]; then
