+++ scripts/tools/publish-nightly.js
@@ -39,7 +39,7 @@ const packagePaths = {
   client: ['client'],
   oauth2: ['plugins', 'auth', 'src', 'auth-oauth2'],
   oidc: ['plugins', 'auth', 'src', 'auth-oidc'],
-  oidc_legacy: ['plugins', 'auth', 'src', 'auth-oidc-legacy'],
+  oidc_pkce: ['plugins', 'auth', 'src', 'auth-oidc-pkce'],
   client_support_angular: ['client-frameworks-support', 'client-support-angular'],
   testing_utilities: ['client-frameworks-support', 'testing-utilities']
 };
@@ -48,7 +48,7 @@ const publishPaths = {
   client: ['client', 'public'],
   oauth2: ['plugins', 'auth', 'public', 'auth-oauth2'],
   oidc: ['plugins', 'auth', 'public', 'auth-oidc'],
-  oidc_legacy: ['plugins', 'auth', 'public', 'auth-oidc-legacy'],
+  oidc_pkce: ['plugins', 'auth', 'public', 'auth-oidc-pkce'],
   client_support_angular: ['client-frameworks-support', 'client-support-angular', 'dist', 'client-support-angular'],
   testing_utilities: ['client-frameworks-support', 'testing-utilities', 'dist']
 };
