+++ test/e2e-test-application/webpack-generateConfig.config.js
@@ -31,12 +31,12 @@ module.exports = {
         },
         // idpProvider OIDC assets
         {
-          from: 'node_modules/@luigi-project/plugin-auth-oidc',
-          to: path.resolve(__dirname, 'src/assets') + '/auth-oidc/'
+          from: 'node_modules/@luigi-project/plugin-auth-oidc-pkce',
+          to: path.resolve(__dirname, 'src/assets') + '/auth-oidc-pkce/'
         },
         {
-          from: 'node_modules/@luigi-project/plugin-auth-oidc-legacy',
-          to: path.resolve(__dirname, 'src/assets') + '/auth-oidc-legacy/'
+          from: 'node_modules/@luigi-project/plugin-auth-oidc',
+          to: path.resolve(__dirname, 'src/assets') + '/auth-oidc/'
         }
       ]
     })
