+++ website/docs/src/markdown-conversion/parser.js
@@ -18,7 +18,7 @@ function setParsedDocs() {
     './../../../../docs',
     './../../../../plugins/auth/public/auth-oauth2',
     './../../../../plugins/auth/public/auth-oidc',
-    './../../../../plugins/auth/public/auth-oidc-legacy'
+    './../../../../plugins/auth/public/auth-oidc-pkce'
   ];
   const parsingArr = [];
   dirs.forEach(dir => {
