+++ website/fiddle/webpack.config.js
@@ -55,7 +55,7 @@ module.exports = {
         { from: './node_modules/@luigi-project/client', to: 'vendor/luigi-client' },
         { from: './node_modules/@luigi-project/plugin-auth-oauth2', to: 'vendor/plugin-auth-oauth2' },
         { from: './node_modules/@luigi-project/plugin-auth-oidc', to: 'vendor/plugin-auth-oidc' },
-        { from: './node_modules/@luigi-project/plugin-auth-oidc-legacy', to: 'vendor/plugin-auth-oidc-legacy' },
+        { from: './node_modules/@luigi-project/plugin-auth-oidc-pkce', to: 'vendor/plugin-auth-oidc-pkce' },
         { from: './node_modules/fundamental-styles', to: 'vendor/fundamental-styles' },
         { from: './node_modules/@sap-theming/theming-base-content', to: 'vendor/theming-base-content' },
         ...['ace.js', 'mode-javascript.js', 'worker-javascript.js', 'theme-textmate.js'].map(f => ({
